+++ NEWS.md
@@ -1,3 +1,9 @@
+## rNodal.PVT 0.0.1.9003
+* 20171115. Bo, Rs Standing
+* add function Rs.Standing. add unit test for example 4.8 Tarek Ahmed.
+* add project book to explain how the project is being built.
+* add CSV file with example from Tarek Ahmed's book. Will use for unit tests.
+
 ## rNodal.PVT 0.0.1.9002
 * 20171114. Add 1st Pb equation. Standing
 * add equation for Bo saturated. add unit test for Bo.